feat: add solution for modules task
